+++ lld/MachO/UnwindInfoSection.cpp
@@ -343,8 +343,7 @@ void UnwindInfoSectionImpl::relocateCompactUnwind(
     if (!d->unwindEntry)
       return;
 
-    // If we have DWARF unwind info, create a slimmed-down CU entry that points
-    // to it.
+    // If we have DWARF unwind info, create a CU entry that points to it.
     if (d->unwindEntry->getName() == section_names::ehFrame) {
       // The unwinder will look for the DWARF entry starting at the hint,
       // assuming the hint points to a valid CFI record start. If it
@@ -361,9 +360,7 @@ void UnwindInfoSectionImpl::relocateCompactUnwind(
       cu.encoding = target->modeDwarfEncoding | dwarfOffsetHint;
       const FDE &fde = cast<ObjFile>(d->getFile())->fdes[d->unwindEntry];
       cu.functionLength = fde.funcLength;
-      // Omit the DWARF personality from compact-unwind entry so that we
-      // don't need to encode it.
-      cu.personality = nullptr;
+      cu.personality = fde.personality;
       cu.lsda = fde.lsda;
       return;
     }