+++ llvm/include/llvm/MC/MCAsmBackend.h
@@ -25,8 +25,7 @@ class MCRelaxableFragment;
 class MCSymbol;
 class MCAsmLayout;
 class MCAssembler;
-class MCContext;
-struct MCDwarfFrameInfo;
+class MCCFIInstruction;
 struct MCFixupKindInfo;
 class MCInst;
 class MCObjectStreamer;
@@ -211,8 +210,8 @@ public:
   virtual void handleAssemblerFlag(MCAssemblerFlag Flag) {}
 
   /// Generate the compact unwind encoding for the CFI instructions.
-  virtual uint32_t generateCompactUnwindEncoding(const MCDwarfFrameInfo *FI,
-                                                 const MCContext *Ctxt) const {
+  virtual uint32_t
+      generateCompactUnwindEncoding(ArrayRef<MCCFIInstruction>) const {
     return 0;
   }
 
@@ -220,8 +219,6 @@ public:
   virtual bool isMicroMips(const MCSymbol *Sym) const {
     return false;
   }
-
-  bool isDarwinCanonicalPersonality(const MCSymbol *Sym) const;
 };
 
 } // end namespace llvm